Fixed printf format and parameters.
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@24514 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/syncml/formatter/Formatter.cpp
@@ -1081,7 +1081,7 @@ StringBuffer* Formatter::getSync(Sync* sync) {
         numberOfChanges->sprintf(
             "<%s>%d</%s>",
             NUMBER_OF_CHANGES,
-            sync->getNumberOfChanges(),
+            (int)sync->getNumberOfChanges(),
             NUMBER_OF_CHANGES);
     }
